perf: preserve console boot module budget
